requirements/base.txt
Django==4.2.*
djangorestframework==3.14.*
django-environ==0.10.*
django-cors-headers
